third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/ceil.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] ceil float16 0D scalar
  assert_less_than_equal: assert_array_approx_equals_ulp: test ceil float16 actual 0 should be close enough to expected 68 by ULP distance: expected a number less than or equal to 0 but got 21568
[FAIL] ceil float16 1D constant tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test ceil float16 actual 0 should be close enough to expected 68 by ULP distance: expected a number less than or equal to 0 but got 21568
[FAIL] ceil float16 1D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test ceil float16 actual 0 should be close enough to expected 68 by ULP distance: expected a number less than or equal to 0 but got 21568
[FAIL] ceil float16 2D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test ceil float16 actual 0 should be close enough to expected 68 by ULP distance: expected a number less than or equal to 0 but got 21568
[FAIL] ceil float16 3D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test ceil float16 actual 0 should be close enough to expected 68 by ULP distance: expected a number less than or equal to 0 but got 21568
[FAIL] ceil float16 4D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test ceil float16 actual 0 should be close enough to expected 68 by ULP distance: expected a number less than or equal to 0 but got 21568
Harness: the test ran to completion.
